naming standartlara göre class moduller düzeltildi
