+++ src/domain/attendance.ts
